requirements.txt
requests==2.31.0
python-dotenv==1.0.1
selenium>=4.10.0